Added refs to BQL
It started to hurt on me that BQL was not mentioned, and in many
ways, that is the most needed part of the fq_codel implementation
on real hardware be it ethernet, dsl, or cable.

I thought about expanding that bit a bit, too.

Also added refs to HTB, HFSC.

I do not like where I stuck a ref to fred's draft. Perhaps it needs a
sentence's explanation somewhere else.

Untested to build (don't have toke's tools)

I also don't mind if my name gets an umlaut.